Don't panic on overallocation and remove mux (#8)

+++ allocator/nonmoving_other.go
@@ -4,7 +4,7 @@ package allocator
 
 import "github.com/tetratelabs/wazero/experimental"
 
-var pageSize = 0 // used only for test
+var pageSize = uint64(0) // used only for test
 
 func alloc(cap, max uint64) experimental.LinearMemory {
 	return sliceAlloc(cap, max)
